combine two lib files
