+++ src/components/LawOfLargeNumbers/LLNSimulation.js
@@ -13,8 +13,9 @@ import ChartContainer from '../ChartContainer.js';
 import SampleSizeInput from '../SampleSizeInput.js';
 import SimulateSamples from './SimulateSamples.js';
 import { Alert } from 'reactstrap';
-import { populationMean, sample, differenceOfMeans } from "../../lib/stats-utils.js";
-import { dataFromDistribution } from '../../lib/data-generation.js';
+import { populationMean, dataFromDistribution } from "../../lib/stats-utils.js";
+import _ from "lodash";
+import { round } from "mathjs";
 
 export default function LLNSimulation({ popType, sampleSize }) {
   const [sampled, setSampled] = useState([]);
@@ -27,16 +28,22 @@ export default function LLNSimulation({ popType, sampleSize }) {
     setStage(1);
     const newPop = dataFromDistribution(popType, sampleSize);
     setPopArray(newPop);
-    const newMean = populationMean(newPop);
-    setPopMean(newMean);
+    setPopMean(populationMean(newPop));
     setSampled([]);
   }, [popType, sampleSize]);
 
   const handleClick = (size) => {
-    const {pop, mue} = sample(size, popArray);
-    setSampled(pop);
-    setSampleMean(mue);
+    const sample = _.sampleSize(popArray, size);
+    setSampled(sample);
+    setSampleMean(populationMean(sample));
     setStage(2);
+  }
+
+  const differenceOfMeans = () => {
+    const diff = (
+      round((popMean - sampleMean) * 100) / 100) === 0 ? 0 :
+      round((popMean - sampleMean) * 100) / 100 || '';
+    return diff;
   }
 
   return (
@@ -49,12 +56,12 @@ export default function LLNSimulation({ popType, sampleSize }) {
           <Alert color="success" style={{ padding: 0, marginTop: '1em' }}>
             Sample Mean: {sampleMean || ''}
             <br/>
-            Difference of Means: {differenceOfMeans(popMean, sampleMean)}
+            Difference of Means: {differenceOfMeans()}
           </Alert>
           <Alert color="info">
             According to the law, the average of the results obtained from a large enough sample should be close to the total average of the population, and will tend to become closer the larger the sample is. Make sure to pick several samples, or see below for a simulation to see the law in action.
           </Alert>
-          <SimulateSamples type={popType} sample={(size) => sample(size, popArray).pop} pop={popMean}/>
+          <SimulateSamples type={popType} sample={(size) => _.sampleSize(popArray, size)} pop={popMean}/>
         </div>
       }
     </Collapsable>
